backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
supabase==1.0.3
pyjwt==2.8.0